backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
python-dotenv==1.0.1

# Task queue
celery==5.4.0
redis==5.1.0

# AI APIs
anthropic==0.40.0
openai==1.50.0
httpx==0.27.0

# YouTube
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Utils
pydantic==2.9.0
aiofiles==24.1.0
Pillow>=10.0.0
imageio-ffmpeg>=0.5.1

# v2.1.0: API 키 AES-GCM 암호화
cryptography>=43.0.0
